In Main.hs, let bindings and happy paths were modified to use more common traversal functions rather than bespoke functions.
In BetaReduce.hs, a major change which had cascading effects was that
the data constructor of LetBinding was no longer imported, so the module
had to be rewritten to replace pattern matching on LetBinding with calls
to exported helper functions. Also, the type signature of
'letExprContainerToFinalContainer' was modified to be more general.
Function 'validateRecursionLetBindingTypes' was replaced with a modified
'validateRecursionLetBindingTypesNew', and then
'validateRecursionLetBindingTypesNew' was renamed to
'validateRecursionLetBindingTypes', effectively making it as if
'validateRecursionLetBindingTypesNew' was deleted and
'validateRecursionLetBindingTypes' was updated. In general, there were
stylistic changes rather than semantic changes.

In CLIToLetExpr.hs, a LetBinding pattern match was replaced with calls
to helper functions and some list pattern matches were changed from
matching on constructor (:) to '[,]' per the suggestion of hlint.

In Print.hs, names of imported functions were changed to match the
updated names in the imported module.

In LetExpr.hs, the export of several data structures were changed from
exporting the type and all data constructors to just the type. To allow
replacing instances of pattern matches in other modules which would no
longer be possible do to changes in export, many helper functions were
created and exported. Some function names were updated.