A new file Container.hs was created and all Container-related code in LetExpr.hs was moved to Container.hs.
In LetExpr.hs, the definition of the data type LetExpr was modified and
functions which interact with it had to be modified accordingly. A Trie
newtype TrieLB was defined, so that other code could conveniently
interface with it and I could avoid or at least reduce exporting data
constructors for things like LetBinding and Var. TrieLB functions for
inserting and retrieving were defined and exported. Some
LetBinding-related functions were defined to make it possible to
manipulate and operate on LetBinding values without deconstructing them
into just the variable or just the value.

In BetaReduce.hs, Trie-related code was replaced with TrieLB
equivalents. All instances of ExprText were replaced with Text. Code was
generally updated to be more legible.

In Main.hs, the happy paths were updated to use newer functions defined
in BetaReduce and LetExpr. Also, code and comments were generally
updated to be more legible.

letexpr.cabal was updated to include Container.hs as a module for the
library.

MyLib.hs was updated to import and export module Container.

In CLIToLetExpr.hs, linearReferenceLetExprParse and
mutualReferenceLetExprParse were refactored into calling smaller scoped
parsing functions, thereby reducing repetition and increasing legibility.

In Print.hs, code was updated to use newer exported functions.